+++ packages/adp-tooling/src/prompts/index.ts
@@ -1,5 +1,7 @@
 export { getPrompts as getPromptsForChangeDataSource } from './change-data-source';
 export { default as ConfigInfoPrompter } from './creation';
+export { getPrompts as getPromptsForNewModel } from './add-new-model';
+export { getPrompts as getPromptsForChangeInbound } from './change-inbound';
 export { getPrompts as getPromptsForAddAnnotationsToOData } from './add-annotations-to-odata';
 export * from './creation/prompt-helpers';
 export { getPrompts as getBasicInfoPrompts } from './creation/basic-questions';