+++ packages/adp-tooling/src/types.ts
@@ -9,7 +9,13 @@ export interface DescriptorVariant {
     reference: string;
     id: string;
     namespace: string;
-    content: object[];
+    content: DescriptorVariantContent[];
+}
+
+export interface DescriptorVariantContent {
+    changeType: string;
+    content: Record<string, unknown>;
+    texts?: string;
 }
 
 export interface ToolsSupport {
@@ -193,12 +199,20 @@ export interface CodeExtChange extends CommonChangeProperties {
     };
 }
 
-export const CUSTOMER_BASE = 'CUSTOMER_BASE';
-
 export const enum TemplateFileName {
     Fragment = 'fragment.xml',
     Controller = 'controller.ejs',
     Annotation = 'annotation.xml'
+}
+
+export const enum FlexLayer {
+    CUSTOMER_BASE = 'CUSTOMER_BASE',
+    VENDOR = 'VENDOR'
+}
+
+export const enum NamespacePrefix {
+    CUSTOMER = 'customer.',
+    EMPTY = ''
 }
 
 export const enum HttpStatusCodes {
@@ -316,30 +330,28 @@ export interface ComponentUsagesData {
 }
 
 export interface NewModelData {
-    projectData: AdpProjectData;
-    timestamp: number;
-    annotation: {
-        /** Name of the OData annotation data source. */
-        dataSourceName: string;
-        /** Optional URI of the OData annotation data source. */
-        dataSourceURI?: string;
-        /** Optional settings for the OData annotation. */
-        settings?: string;
-    };
-    service: {
-        /** Name of the OData service. */
-        name: string;
-        /** URI of the OData service. */
-        uri: string;
-        /** Name of the OData service model. */
-        modelName: string;
-        /** Version of OData used. */
-        version: string;
-        /** Settings for the OData service model. */
-        modelSettings: string;
-    };
-    /** Indicates whether annotation mode is added. */
+    variant: DescriptorVariant;
+    answers: NewModelAnswers;
+}
+
+export interface NewModelAnswers {
     addAnnotationMode: boolean;
+    /** Name of the OData service. */
+    name: string;
+    /** URI of the OData service. */
+    uri: string;
+    /** Name of the OData service model. */
+    modelName: string;
+    /** Version of OData used. */
+    version: string;
+    /** Settings for the OData service model. */
+    modelSettings: string;
+    /** Name of the OData annotation data source. */
+    dataSourceName: string;
+    /** Optional URI of the OData annotation data source. */
+    dataSourceURI?: string;
+    /** Optional settings for the OData annotation. */
+    annotationSettings?: string;
 }
 
 export interface DataSourceData {
@@ -348,21 +360,20 @@ export interface DataSourceData {
     answers: ChangeDataSourceAnswers;
 }
 
+export interface InboundChangeAnswers {
+    /** Title associated with the inbound navigation data. */
+    title: string;
+    /** Subtitle associated with the inbound navigation data. */
+    subTitle: string;
+    /** Icon associated with the inbound navigation data. */
+    icon: string;
+}
+
 export interface InboundData {
-    projectData: AdpProjectData;
-    timestamp: number;
     /** Identifier for the inbound navigation data. */
     inboundId: string;
-    flp: {
-        /** Title associated with the inbound navigation data. */
-        title: PropertyValueType;
-        /** Subtitle associated with the inbound navigation data. */
-        subTitle: PropertyValueType;
-        /** Icon associated with the inbound navigation data. */
-        icon: PropertyValueType;
-    };
-    /** Optional flag indicating if the project is in safe mode. */
-    isInSafeMode?: boolean;
+    variant: DescriptorVariant;
+    answers: InboundChangeAnswers;
 }
 
 export interface InboundContent {