+++ packages/inquirer-common/src/types.ts
@@ -96,6 +96,11 @@ export interface ConfirmQuestion<A extends Answers = Answers> extends BaseConfir
     guiOptions?: YUIQuestion['guiOptions'];
 }
 
+export interface EditorQuestion<A extends Answers = Answers> extends BaseEditorQuestion<A> {
+    name: YUIQuestion['name'];
+    guiOptions?: YUIQuestion['guiOptions'];
+}
+
 export interface InputQuestion<A extends Answers = Answers> extends BaseInputQuestion<A> {
     name: YUIQuestion['name'];
     guiOptions?: YUIQuestion['guiOptions'];